Rename SOR entry point
